pyproject.toml
[project]
name = "axion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Aleš Kučera", email = "kuceral4@fel.cvut.cz" }
]
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.3",
    "pytest>=8.4.0",
    "torch>=2.7.1",
    "warp-lang>=1.7.2",
]

[project.scripts]
axion = "axion:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
